# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Interactive Fiction Platform API
  description: OpenAPI 3.0 Specification for Interactive Fiction Platform (DAG Novel System)
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local Development Server
paths:
...
  /api/register:
    post:
      summary: Register new user
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [username, email, password]
              properties:
                username:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                  description: "ขั้นต่ำ 8 ตัวอักษร ต้องมีตัวอักษรและตัวเลขอย่างละอย่างน้อย 1 ตัว"
                profileImage:
                  type: string
                  format: binary
                  description: "รูปโปรไฟล์ผู้ใช้ (ไม่บังคับ)"
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/login:
    post:
      summary: User login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  refresh_token:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/refresh:
    post:
      summary: Refresh authentication token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/users:
    get:
      summary: Get current user profile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/me/novels:
    get:
      summary: Get novels authored by current user (Writer)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of writer novels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Novel'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/me/username:
    patch:
      summary: Update current user's username
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username]
              properties:
                username:
                  type: string
      responses:
        '200':
          description: Username updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/me/email:
    patch:
      summary: Update current user's email
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Email updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/me/password:
    patch:
      summary: Change current user's password
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password, confirm_password]
              properties:
                current_password:
                  type: string
                new_password:
                  type: string
                confirm_password:
                  type: string
      responses:
        '200':
          description: Password updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/me/profile-picture:
    patch:
      summary: Update current user's profile picture
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [profileImage]
              properties:
                profileImage:
                  type: string
                  format: binary
      responses:
        '200':
          description: Profile picture updated
          content:
            application/json:
              schema:
                type: object
                required: [message, pic_profile]
                properties:
                  message:
                    type: string
                  pic_profile:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      summary: Update current user's profile picture
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [profileImage]
              properties:
                profileImage:
                  type: string
                  format: binary
      responses:
        '200':
          description: Profile picture updated
          content:
            application/json:
              schema:
                type: object
                required: [message, pic_profile]
                properties:
                  message:
                    type: string
                  pic_profile:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/me:
    delete:
      summary: Delete current user's account
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password]
              properties:
                current_password:
                  type: string
      responses:
        '200':
          description: Account deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/me/suspend:
    patch:
      summary: Suspend current user's account
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Account suspended
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/me/notification-settings:
    get:
      summary: Get current user's notification settings
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Notification settings
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/NotificationSettings'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      summary: Update current user's notification settings
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationSettings'
      responses:
        '200':
          description: Notification settings updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      summary: Update current user's notification settings
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationSettings'
      responses:
        '200':
          description: Notification settings updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /novels:
    get:
      summary: List all novels
      responses:
        '200':
          description: List of novels
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Novel'
    post:
      summary: Create new novel (Writer)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title:
                  type: string
                captions:
                  type: string
                  nullable: true
                introduction:
                  type: string
                  nullable: true
                cover_image:
                  type: string
                  nullable: true
                category_ids:
                  type: array
                  items:
                    type: integer
                status:
                  type: string
                is_published:
                  type: boolean
                is_completed:
                  type: boolean
      responses:
        '201':
          description: Novel created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [message, novel_id]
                        properties:
                          message:
                            type: string
                          novel_id:
                            type: integer
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /choices:
    post:
      summary: Create choice connecting from_scene to to_scene (DAG Edge)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [from_scene_id, to_scene_id, label]
              properties:
                from_scene_id:
                  type: integer
                to_scene_id:
                  type: integer
                label:
                  type: string
      responses:
        '201':
          description: Choice created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [message, choice_id]
                        properties:
                          message:
                            type: string
                          choice_id:
                            type: integer
        '400':
          description: Bad request or DAG loop conflict detected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          description: Method Not Allowed
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad Request Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource Not Found Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Resource Conflict Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    SuccessResponse:
      type: object
      required: [status, message]
      properties:
        status:
          type: integer
        message:
          type: string
        data:
          nullable: true
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: BAD_REQUEST
            message:
              type: string
              example: Invalid input payload
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [reader, writer, admin]
...
    Novel:
      type: object
      properties:
        novel_id:
          type: integer
        title:
          type: string
        captions:
          type: string
        introduction:
          type: string
        cover_image:
          type: string
        status:
          type: string
          enum: [draft, published, suspended]
        is_published:
          type: boolean
        is_completed:
          type: boolean
        is_banned:
          type: boolean
        ban_reason:
          type: string
        author_id:
          type: integer
        author_name:
          type: string
        pen_name:
          type: string
        views:
          type: integer
        chapter_count:
          type: integer
        scene_count:
          type: integer
        like_count:
          type: integer
        bookshelf_count:
          type: integer
...
    NotificationSettings:
      type: object
      required: [in_app_notifications, novel_updates, comments, likes, follows]
      properties:
        in_app_notifications:
          type: boolean
        novel_updates:
          type: boolean
        comments:
          type: boolean
        likes:
          type: boolean
        follows:
          type: boolean
    AuthResponse:
      type: object
      description: "Response จาก Login และ Register"
      properties:
        token:
          type: string
          description: "JWT Access Token"
        refresh_token:
          type: string
          description: "JWT Refresh Token"
        user:
          type: object
          properties:
            id:
              type: integer
            username:
              type: string
            email:
              type: string
            pic_profile:
              type: string
            role:
              type: string
              enum: [reader, writer, admin]
...